refactor(backend): removed internal api, moved web
