--- new/src/backend/app/api/v1/routes/web/protonation.py
@@ -1,5 +1,0 @@
-"""Routes for protonation."""
-
-from fastapi.routing import APIRouter
-
-protonation_router = APIRouter(prefix="/protonation")
