--- new/src/backend/app/api/v1/routes/web/user.py
@@ -1,5 +1,0 @@
-"""Routes for user management."""
-
-from fastapi.routing import APIRouter
-
-user_router = APIRouter(prefix="/users")
