+++ new/src/backend/app/core/logging/file_logger.py
@@ -22,9 +22,7 @@ class FileLogger(LoggerBase):
             os.makedirs(self.logdir)
 
         logging.basicConfig(
-            filename=os.path.join(
-                self.logdir, self.file_name
-            ),  # TODO: use separate log files for web and internal apis?
+            filename=os.path.join(self.logdir, self.file_name),
             level=logging.INFO,
             filemode="a+",
             format="%(asctime)s [%(levelname)s] %(message)s",  # TODO: what format should be used?
